fix: cover cases for filters
